notebook schema
