+++ src/app/virtual-lab/lab/[virtualLabId]/project/[projectId]/(pages)/notebooks/member/page.tsx
@@ -1,20 +1,10 @@
-import { z } from 'zod';
+import { NotebooksArraySchema } from '../schemas';
 import UserNotebookPage from './UserNotebookPage';
 import { ServerSideComponentProp } from '@/types/common';
 import { fetchNotebook } from '@/util/virtual-lab/github';
 import { virtualLabApi } from '@/config';
 import { assertVLApiResponse } from '@/util/utils';
 import authFetch from '@/authFetch';
-
-const NotebookSchema = z.object({
-  id: z.string().uuid(),
-  project_id: z.string().uuid(),
-  github_file_url: z.string().url(),
-  created_at: z.string().datetime(),
-  updated_at: z.string().datetime(),
-});
-
-const NotebooksArraySchema = z.array(NotebookSchema);
 
 export default async function Notebooks({
   params,
